+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelogs
 
-## November 9, 2016 (version 1.2.3)
+## November 9, 2016 (version 1.2.5)
+  - ADD: Hints of the synchronization.
+  - ADD: Reload dialog after extensions are synced.
   - ADD: Support download settings from public Gist.
   - FIX: Synced extensions aren't managed by vscode.